Fix bug with VLSub '401 Unauthorized' error
Using the 'LogIn' XML-RPC with a none-English language results in no
subtitles being listed and an error message: "Error: code '401
Unauthorized (200)'".

Problem:
'LogIn' RPC is requested with the 'language' as the 'useragent'
parameter,
api.opensubtitles.org replies with '414 Unkown User Agent' error
instead of the expected session token,
'SearchSubtitles' RPC is requested with an empty session token,
api.opensubtitles.org replies with '401 Unauthorised' error instead
of the expected listing of subtitles.

Solution: reorder the parameters in the 'LogIn' call to meet with the API
documented in:
https://trac.opensubtitles.org/projects/opensubtitles/wiki/XMLRPC#LogIn
which specifies: LogIn( $username, $password, $language, $useragent )
Notice how $language and $useragent were out of order.

Steps to reproduce bug:
1) Open any video file with vlc
2) Open VLSub extension
3) Select 'Hebrew' from the 'Subtitles language:' dropdown menu
4) Click 'Search by hash' or 'Search by name'